% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\description{
This script loads and aggregates all data for the project.
Data are stored in base_dir/data
There are three directories in the data folder
\enumerate{
\item codebooks - store all excel sheets. These are updated by the study team
\item redcap - the redcap project will be exported to this folder
\item output - any output data will be stored in the output folder \cr \cr
}
}
\section{Development}{

5.9.22 Began development. JC \cr
5.10.22 Initial build complete. JC \cr
}

\author{
Jon Clutton
}
