% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_script.R
\name{master_script}
\alias{master_script}
\title{Master Script of vCCCAudit Project}
\description{
This is the master script for vCCC_Audit.
It is written currently written to run manually. However, it is designed
to switch onto the shiny server.
Libraries, directories, and scripts to be run are all declared in this
script. The variable no_email can be changed to 1 to test the project and
stop emails.
}
\section{Instructions to Run}{

Download the current REDCap database and save to data/redcap. \cr
Press Source \cr
}

\section{Directory Shortcuts}{

base_dir - project level directory \cr
data_dir - data \cr
script_dir - R scripts \cr
module_dir - random python and .txt files associated with projects \cr
codebook_dir - within data_dir, codebooks \cr
output_dir - within data_dir, output data \cr
}

\section{Project Level Variables}{

no_email - stops emails if set to any number other than 0 \cr
monitor_fname \cr
monitor_lname \cr
monitor_email \cr
}

\section{Development}{

9.27.22 Began Project JC \cr
9.27.22 Initial build complete. JC \cr
9.28.22 Reviewed with Heidi and Amber. Making changes to emails and REDCap. JC \cr
10.7.22 Completing final testing and documenting. Plan to initiate next week. JC \cr
}

\author{
Jon Clutton
}
