% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_emails.R
\name{send_emails}
\alias{send_emails}
\title{send_emails}
\description{
An html table is created as the body of the email and written to modules/send_staff_time_email_module. The email is sent using
python. To change senders or recipients of the email, change manually within the python script found here - modules/send_staff_time_email_module/send_email.py \cr
The purpose of this script is to send individual emails to all raters in the vCCC project
}
\section{Development}{

9.27.22 Began development. JC \cr
9.27.22 Initial build complete. JC \cr
}

\author{
Jon Clutton
}
