% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_monitor_email.R
\name{send_monitor_email}
\alias{send_monitor_email}
\title{send_monitor_email}
\description{
An html table is created as the body of the email and written to modules/send_monitor_email_module. The email is sent using
python. To change senders or recipients of the email, change manually within the python script found here - modules/send_monitor_email_module/send_email.py \cr
The purpose of this script is to send the study monitor a review of all ongoing open inquiries in the vCCC monitoring project
}
\section{Development}{

10.5.22 Began development. JC \cr
}

\author{
Jon Clutton
}
